requirements.txt
<<<<<<< HEAD
pandas
numpy
requests
beautifulsoup4
newspaper3k
feedparser
regex
spacy
nltk
scikit-learn
vaderSentiment
bertopic
sentence-transformers
umap-learn
hdbscan
matplotlib
plotly
streamlit
pydantic
SQLAlchemy
python-dotenv
tqdm
newsapi-python
textblob

=======
pandas
numpy
requests
beautifulsoup4
newspaper3k
feedparser
regex
spacy
nltk
scikit-learn
vaderSentiment
bertopic
sentence-transformers
umap-learn
hdbscan
matplotlib
plotly
streamlit
pydantic
SQLAlchemy
python-dotenv
tqdm
newsapi-python
textblob 
>>>>>>> f140fda (Initial commit: full project structure)
